requirements.txt
absl-py==2.3.0
anyio==3.7.1
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
audioread==3.0.1
better-profanity==0.7.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
demucs==4.0.1
exceptiongroup==1.3.0
executing==2.2.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h11==0.12.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
langdetect==1.0.9
libclang==18.1.1
librosa>=0.9.0
llvmlite>=0.39.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.2.1
norbert==0.2.1
numba>=0.55.0
numpy==1.26.4
openai-whisper==20240930
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
resampy==0.3.1
rfc3986==1.5.0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
sniffio==1.3.1
soundfile==0.13.1
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
whitenoise
